requirements.txt
easyocr==1.7.2
faiss-cpu==1.13.0
faster-whisper==1.2.1
groq==0.37.1
langchain==0.3.28
langchain-community==0.3.31
langchain-core==0.3.83
langchain-groq==0.3.8
langchain-text-splitters==0.3.11
langgraph==0.6.11
langgraph-checkpoint==2.1.2
langgraph-prebuilt==0.6.5
langgraph-sdk==0.2.9
opencv-python==4.13.0.92
opencv-python-headless==4.13.0.92
pillow==11.3.0
pydantic==2.12.5
pydantic-settings==2.11.0
pydantic_core==2.41.5
python-dotenv==1.2.1
sentence-transformers==5.1.2
streamlit==1.50.0
